test: shadow to_timestamp(double precision) in bigint migration case
The emitted USING expression passes float8 to to_timestamp, so plant a
shadow on that overload to catch regressions if pg_catalog qualification
is removed.

Committed-By-Agent: cursor

+++ internal/migration_acceptance_tests/column_cases_test.go
@@ -312,7 +312,7 @@ var columnAcceptanceTestCases = []acceptanceTestCase{
 		name: "Change BIGINT to TIMESTAMP (ignores shadowed to_timestamp in USING)",
 		oldSchemaDDL: []string{
 			`
-            CREATE FUNCTION public.to_timestamp(bigint)
+            CREATE FUNCTION public.to_timestamp(double precision)
             RETURNS timestamp without time zone LANGUAGE plpgsql AS $$
             BEGIN
               RAISE EXCEPTION 'pwnd';
@@ -327,7 +327,7 @@ var columnAcceptanceTestCases = []acceptanceTestCase{
 		},
 		newSchemaDDL: []string{
 			`
-            CREATE FUNCTION public.to_timestamp(bigint)
+            CREATE FUNCTION public.to_timestamp(double precision)
             RETURNS timestamp without time zone LANGUAGE plpgsql AS $$
             BEGIN
               RAISE EXCEPTION 'pwnd';
